:bug: Fix foreign key in vehicle component table

+++ database/migrations/base_structure/api/starcitizen/vehicle/vehicle/2020_09_18_124902_create_vehicle_component_table.php
@@ -24,8 +24,8 @@ class CreateVehicleComponentTable extends Migration
         Schema::create(
             'vehicle_component',
             function (Blueprint $table) {
-                $table->unsignedInteger('vehicle_id');
-                $table->unsignedInteger('component_id');
+                $table->unsignedBigInteger('vehicle_id');
+                $table->unsignedBigInteger('component_id');
 
                 $table->foreign('vehicle_id')
                     ->references('id')
